Rename models. Client side
