+++ src/StickedWords.UI.React/src/models/exercises/Guess.ts
@@ -1,4 +1,4 @@
-export class TranslateGuess {
+export class Guess {
   flashCardId!: number;
   answer!: string;
 }
