+++ src/StickedWords.UI.React/src/models/exercises/TranslateExercise.ts
